Group List with admin name Committed

+++ routes/api.php
@@ -27,24 +27,11 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
 
-Route::get('/email', [AdminController::class, 'mail']);
-
-Route::post('/login', [AuthController::class, 'login']);
-Route::post('/logout', [AuthController::class, 'logout']);
-Route::post('/ttt', [AuthController::class, 'ttt']);
 Route::get('/sms', [SmsController::class, 'aws']);
-
- //Route::group(['middleware' => 'auth:sanctum',], function (){
- Route::get('/test', [NotificationController::class, 'twilioSms']);
- //});
-
-// Route::get('/meow', [TestController::class, 'meow'])->middleware('admin');
+Route::get('/test', [NotificationController::class, 'twilioSms']);
 
 
 Route::group(['middleware' => 'auth:sanctum',], function (){
-
-    // Route::get('/meow', [TestController::class, 'meow'])->middleware('admin');
-
     Route::get('/admin', [AdminController::class, 'index'])->middleware('super_admin');
     Route::post('/admin', [AdminController::class, 'store'])->middleware('super_admin');
     Route::get('/admin/{id}', [AdminController::class, 'show'])->middleware('super_admin');
@@ -53,33 +40,35 @@ Route::group(['middleware' => 'auth:sanctum',], function (){
 
     Route::get('/users-count', [AdminController::class, 'profileCount'])->middleware('super_admin');
     Route::get('/admin-count', [AdminController::class, 'count'])->middleware('super_admin');
-    Route::post('/createadmin', [AdminController::class, 'createadmin'])->middleware('super_admin');
+    Route::post('/create-admin', [AdminController::class, 'createAdmin'])->middleware('super_admin');
     Route::get('/birthdays', [BirthdayController::class, 'birthdays'])->middleware('super_admin');
     Route::get('/birthdays-count', [BirthdayController::class, 'birthdayscount'])->middleware('super_admin');
-
 });
 
 Route::group(['middleware' => 'auth:sanctum',], function (){
 
-    Route::get('/group', [GroupController::class, 'index']);
-    Route::post('/group', [GroupController::class, 'store']);
-    Route::get('/group/{id}', [GroupController::class, 'show']);
-    Route::put('/group/{id}', [GroupController::class, 'update']);
+    Route::get('/group', [GroupController::class, 'index'])->middleware('super_admin');
+    Route::post('/group', [GroupController::class, 'store'])->middleware('super_admin');
+    Route::get('/group/{id}', [GroupController::class, 'show'])->middleware('super_admin');
+    Route::post('/group/{id}', [GroupController::class, 'update'])->middleware('super_admin');
 
-    Route::get('/group-count', [GroupController::class, 'count']);
+    Route::get('/group-count', [GroupController::class, 'count'])->middleware('super_admin');
 });
-
 
 Route::group(['middleware' => 'auth:sanctum',], function (){
-    Route::get('/user', [UserController::class, 'index']);
-    Route::post('/user', [UserController::class, 'store']);
-    Route::delete('/user/{id}', [UserController::class, 'destroy']);
-    Route::put('/user/{id}', [UserController::class, 'update']);
-    Route::get('/user/{id}', [UserController::class, 'show']);
+    Route::get('/user', [UserController::class, 'index'])->middleware('admin');
+    Route::post('/user', [UserController::class, 'store'])->middleware('admin');
+    Route::delete('/user/{id}', [UserController::class, 'destroy'])->middleware('admin');
+    Route::post('/user/{id}', [UserController::class, 'update'])->middleware('admin');
+    Route::get('/user/{id}', [UserController::class, 'show'])->middleware('admin');
 
-    Route::get('/user-count', [UserController::class, 'count']);
-    Route::get('/birthday-count', [BirthdayController::class, 'birthdaycount']);
-    Route::get('/birthday', [BirthdayController::class, 'birthday']);
-
-    //Route::get('/test', [TestController::class, 'index']);
+    Route::get('/user-count', [UserController::class, 'count'])->middleware('admin');
+    Route::get('/birthday-count', [BirthdayController::class, 'birthdaycount'])->middleware('admin');
+    Route::get('/birthday', [BirthdayController::class, 'birthday'])->middleware('admin');
 });
+
+Route::post('/login', [AuthController::class, 'login']);
+Route::post('/logout', [AuthController::class, 'logout']);
+
+
+
